Make compatible with CaloGeometry changes

+++ Geometry/EcalTestBeam/plugins/EcalTBHodoscopeGeometryLoaderFromDDD.cc
@@ -85,15 +85,15 @@ void EcalTBHodoscopeGeometryLoaderFromDDD::makeGeometry(const DDCompactView* cpv
       const GlobalPoint refPoint ( ctr.x(), ctr.y(), ctr.z() ) ;
 
 
-      std::vector<float> vv ;
+      std::vector<double> vv ;
       vv.reserve( pv.size() ) ;
       for( unsigned int i ( 0 ) ; i != pv.size() ; ++i )
       {
 	 vv.push_back( CaloCellGeometry::k_ScaleFromDDDtoGeant*pv[i] ) ;
       }
-      const float* pP ( CaloCellGeometry::getParmPtr( vv, 
-						      ebg->parMgr(), 
-						      ebg->parVecVec() ) ) ;
+      const double* pP ( CaloCellGeometry::getParmPtr( vv, 
+						       ebg->parMgr(), 
+						       ebg->parVecVec() ) ) ;
 
       PreshowerStrip* cell ( new PreshowerStrip( refPoint,
 						 ebg->cornersMgr(),
